refactor: clean code improvements
- Fix README: correct test:mcp script reference and .user-data/ directory name
- Deduplicate escapeHtml: share escapeHtmlChars from parsers, remove duplicate in chatsvc-messaging
- Add withLocalStorage wrapper to reduce boilerplate in token-extractor
- Split parsers.ts (1394 lines) into 7 focused submodules with barrel re-exports
- Parallelize unread checks in message-tools using Promise.allSettled
- Eliminate eslint-disable in registry.ts with proper RegistryEntry typing
- Remove deprecated buildMessageLink positional overload
- Add lightweight logger abstraction replacing console.error/log calls
- Replace IIFE timestamp fallback with named timestampFromIdOrNow helper

All 225 tests pass, TypeScript compiles cleanly.
