+++ src/utils/parsers-html.ts
@@ -1,0 +1,56 @@
+/**
+ * HTML parsing and escaping utilities.
+ */
+
+import type { ExtractedLink } from '../types/teams.js';
+
+// Re-export ExtractedLink so existing imports from parsers.ts continue to work
+export type { ExtractedLink };
+
+/**
+ * Extracts links from HTML content before stripping.
+ * Returns an array of { url, text } objects.
+ */
+export function extractLinks(html: string): ExtractedLink[] {
+  const links: ExtractedLink[] = [];
+  const linkRegex = /<a\s+[^>]*href=["']([^"']+)["'][^>]*>([\s\S]*?)<\/a>/gi;
+  
+  let match;
+  while ((match = linkRegex.exec(html)) !== null) {
+    const url = match[1];
+    const text = stripHtml(match[2]); // Clean nested HTML in link text
+    if (url && !url.startsWith('javascript:')) {
+      links.push({ url, text: text || url });
+    }
+  }
+  
+  return links;
+}
+
+/**
+ * Strips HTML tags from content for display.
+ */
+export function stripHtml(html: string): string {
+  return html
+    .replace(/<[^>]*>/g, ' ')
+    .replace(/&nbsp;/g, ' ')
+    .replace(/&amp;/g, '&')
+    .replace(/&lt;/g, '<')
+    .replace(/&gt;/g, '>')
+    .replace(/&quot;/g, '"')
+    .replace(/&#39;/g, "'")
+    .replace(/&apos;/g, "'")
+    .replace(/\s+/g, ' ')
+    .trim();
+}
+
+/**
+ * Escapes HTML special characters in text.
+ */
+export function escapeHtmlChars(text: string): string {
+  return text
+    .replace(/&/g, '&amp;')
+    .replace(/</g, '&lt;')
+    .replace(/>/g, '&gt;')
+    .replace(/"/g, '&quot;');
+}
